reorganize RecoTracker/TrackProducer/plugins/ to avoid SealModules.cc
