+++ RecoTracker/TrackProducer/plugins/TrackProducer.cc
@@ -1,17 +1,53 @@
+/** \class TrackProducer
+ *  Produce Tracks from TrackCandidates
+ *
+ *  \author cerati
+ */
+
 // system include files
 #include <memory>
 
 // user include files
 #include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "Geometry/Records/interface/TrackerTopologyRcd.h"
 #include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
-#include "RecoTracker/TrackProducer/plugins/TrackProducer.h"
+#include "RecoTracker/TrackProducer/interface/KfTrackProducerBase.h"
+#include "RecoTracker/TrackProducer/interface/TrackProducerAlgorithm.h"
 #include "TrackingTools/GeomPropagators/interface/Propagator.h"
 #include "TrackingTools/PatternTools/interface/TrajTrackAssociation.h"
 #include "TrackingTools/PatternTools/interface/Trajectory.h"
+#include "TrackingTools/TransientTrack/interface/TransientTrack.h"
+
+class TrackProducer : public KfTrackProducerBase, public edm::stream::EDProducer<> {
+public:
+  /// Constructor
+  explicit TrackProducer(const edm::ParameterSet& iConfig);
+
+  /// Implementation of produce method
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+  /// Get Transient Tracks
+  std::vector<reco::TransientTrack> getTransient(edm::Event&, const edm::EventSetup&);
+
+  //   /// Put produced collections in the event
+  //   virtual void putInEvt(edm::Event&,
+  // 			std::unique_ptr<TrackingRecHitCollection>&,
+  // 			std::unique_ptr<TrackCollection>&,
+  // 			std::unique_ptr<reco::TrackExtraCollection>&,
+  // 			std::unique_ptr<std::vector<Trajectory> >&,
+  // 			AlgoProductCollection&);
+
+  /// fillDescriptions
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  TrackProducerAlgorithm<reco::Track> theAlgo;
+  edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> theTTopoToken;
+};
 
 TrackProducer::TrackProducer(const edm::ParameterSet& iConfig)
     : KfTrackProducerBase(iConfig.getParameter<bool>("TrajectoryInEvent"),
@@ -177,3 +213,6 @@ std::vector<reco::TransientTrack> TrackProducer::getTransient(edm::Event& theEve
 
   return ttks;
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TrackProducer);
