--- RecoTracker/TrackProducer/plugins/TrackRefitter.h
@@ -1,34 +1,0 @@
-#ifndef TrackRefitter_h
-#define TrackRefitter_h
-
-/** \class TrackRefitter
- *  Refit Tracks: Produce Tracks from TrackCollection. It performs a new final fit on a TrackCollection.
- *
- *  \author cerati
- */
-
-#include "FWCore/Framework/interface/stream/EDProducer.h"
-#include "RecoTracker/TrackProducer/interface/KfTrackProducerBase.h"
-#include "RecoTracker/TrackProducer/interface/TrackProducerAlgorithm.h"
-
-class TrackRefitter : public KfTrackProducerBase, public edm::stream::EDProducer<> {
-public:
-  /// Constructor
-  explicit TrackRefitter(const edm::ParameterSet& iConfig);
-
-  /// Implementation of produce method
-  void produce(edm::Event&, const edm::EventSetup&) override;
-
-  /// fillDescriptions
-  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
-
-private:
-  TrackProducerAlgorithm<reco::Track> theAlgo;
-  enum Constraint { none, momentum, vertex, trackParameters };
-  Constraint constraint_;
-  edm::EDGetToken trkconstrcoll_;
-
-  const edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> ttopoToken_;
-};
-
-#endif
